test: reorganization of the 'flush' unit test
