+++ tests/unit/CMakeLists.txt
@@ -7,7 +7,7 @@ add_subdirectory(conn)
 add_subdirectory(conn_req)
 add_subdirectory(ep)
 add_subdirectory(error)
-add_subdirectory(flush-test)
+add_subdirectory(flush)
 add_subdirectory(info)
 add_subdirectory(log)
 add_subdirectory(mr)
