+++ tests/unit/common/mocks-stdlib.h
@@ -1,0 +1,17 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * Copyright 2020, Intel Corporation
+ */
+
+/*
+ * mocks-stdlib.h -- the stdlib mocks' header
+ */
+
+#ifndef MOCKS_STDLIB_H
+#define MOCKS_STDLIB_H
+
+struct posix_memalign_args {
+	void *ptr;
+};
+
+#endif /* MOCKS_STDLIB_H */
