+++ tests/unit/common/mocks-unistd.h
@@ -1,0 +1,15 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * Copyright 2020, Intel Corporation
+ */
+
+/*
+ * mocks-unistd.h -- the unistd mocks' header
+ */
+
+#ifndef MOCKS_UNISTD_H
+#define MOCKS_UNISTD_H
+
+#define PAGESIZE 4096
+
+#endif /* MOCKS_UNISTD_H */
